+++ src/util.js
@@ -15,3 +15,26 @@ export const getUniqueItem = (items) => {
   return b;
 };
 
+export const getTimeInfo = (dateEnd, dateStart) => {
+  let timeStr = ``;
+  let day;
+  let hour;
+  let minute = dateEnd.diff(dateStart, `m`);
+  if (minute < 0) {
+    minute = minute * -1;
+  }
+  if (minute > 1440) {
+    day = Math.floor(minute / 1440);
+    hour = Math.floor((minute - day * 1440) / 60);
+    minute = minute - ((minute - day * 1440) % 60);
+    timeStr = `${day}D ${hour}H ${minute}M`;
+  } else if (minute >= 60 && minute < 1440) {
+    hour = Math.floor(minute / 60);
+    minute = minute - hour * 60;
+    timeStr = `${hour}H ${minute}M`;
+  } else if (minute > 0 && minute < 60) {
+    minute = minute;
+    timeStr = `${minute}M`;
+  }
+  return timeStr;
+};
